Fix test dependencies emitted by the recorder
For reading, use the Gradle model (not the parser) for checking
dependencies. This way we check the effective dependency graph, as seen
by Gradle and obtained at sync time. This means we understand transitive
dependencies and any other magic tricks that the parser does not
support.

For writing emit androidTestCompile dependencies with exclude blocks.

Bug: http://b.android.com/218935
Change-Id: Id51c20eeb92633433eb7c0e228c4ccbb803be1e1